HEader & Footer components moved to specific folders
