+++ app/layout.js
@@ -1,7 +1,7 @@
 import "./globals.css";
 
-import Footer from "./components/Footer";
-import Header from "./components/Header";
+import Footer from "./components/footer/Footer";
+import Header from "./components/header/Header";
 
 export const metadata = {
   title: "Portfolio Website",
